fix: 익명 게시판 요청 & 응답 데이터 수정
* fix: imageUrl, description 필드 추가 및 usedBoard 삭제

* fix: boardId, postId 필드 추가

* refactor: createAt -> createdAt

* fix: member 정보 author로 구조화(nickname 포함)

* fix: boardId, boardTitle, postId 필드 추가

* fix: 익명일때 author 필드 nickname만 값 넣는걸로 수정

* fix: 댓글 member정보 구조화

* bug: nextCursor 마지막값 가져오는 로직 중 널포인터 예외 버그 해결

* fix: anonymous -> anonymity로 수정

* refactor: 닉네임 가져오는 로직 리팩토링

* fix: 게시글 좋아요 응답 데이터 likeCount, liked 추가

* fix: 게시글 스크랩 응답 데이터 scrapCount, scraped 추가

* fix: 게시글 삭제 시 s3 이미지 삭제 로직 추가

* fix: memberId 삭제

* fix: commentId필드 추가, 삭제된 댓글이면 빈 문자열 응답

+++ src/main/java/com/ssafy/ssafsound/domain/post/repository/PostLikeRepository.java
@@ -9,5 +9,6 @@ import java.util.Optional;
 @Repository
 public interface PostLikeRepository extends JpaRepository<PostLike, Long> {
     Optional<PostLike> findByPostIdAndMemberId(Long postId, Long memberId);
-    Long countByPostId(Long postId);
+
+    Integer countByPostId(Long postId);
 }
